# docs/openapi.yaml
openapi: 3.0.0
info:
  title: SaaS Factory Blocks API
  description: API dokumentacija za sve blokove - Auth, Database, Payments, Emails
  version: 0.1.0
  contact:
    name: SaaS Factory
    url: https://github.com/your-org/saas-fact
servers:
  - url: http://localhost:3000
    description: Development
  - url: https://api.yourapp.com
    description: Production
...
paths:
...
  /api/payments/checkout:
    post:
      tags:
        - Payments
      summary: Kreiraj Stripe checkout sesiju
      operationId: createCheckoutSession
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                priceId:
                  type: string
                  description: Stripe price ID (monthly/yearly)
              required:
                - priceId
      responses:
        '200':
          description: Checkout sesija kreirjena
          content:
            application/json:
              schema:
                type: object
                properties:
                  checkoutUrl:
                    type: string
                    format: uri
        '401':
          description: Neautentificiran korisnik
        '404':
          description: Price ne postoji
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token iz Supabase Auth